# Optimize IOperation.Children Implementation