… src/Compilers/Core/Portable/Generated/Operations.Generated.cs
… src/Compilers/Core/Portable/Operations/OperationWalker.cs
… src/Tools/Source/CompilerGeneratorTools/Source/IOperationGenerator/IOperationClassWriter.cs