Optimizes the implemention of IOperation.Children by using a struct enumerable/enumerator internally, with slot/index tracking, and exposing this as ChildOperations on Operation internally. Uses of IOperation.Children in internal compiler code such as Descendents() and OperationWalker have been updated to use this new property, making significant improvements to memory usage. I've taken some benchmarks with 3 different scenarios:

1. master (2766389028b)
2. old-children. This uses the same Children implementation as master, but otherwise uses the new non-lazy generation in features/ioperation.
3. better-children. This is the implementation of Children in this branch.

### AnalyzerRunner Results

Note: times are somewhat variable. I've run the benchmarks a few times and while on the whole ms taken does appear to be improved, I can't say it has with statistic certainty. Memory usage, on the other hand, is much more deterministic, and I can confidently say that we save a few GBs of allocations. This run was performed using Microsoft.CodeAnalysis.Analyzers (which is heavily IOperation-based).

**branch**|** time**|** bytes allocated**
:-----:|:-----:|:-----:
master (2766389028b)| 536240ms|123,258,378,656
old-children| 538872ms|122,012,083,264
better-children| 526784ms|119,468,190,816

### dotnet/performance Roslyn Benchmarks

The main benchmark to pay attention to here is GetDiagnosticsWithAnalyzers, but I included the rest for completeness. There's not a clear-cut win in terms of execution times, but there is nearly 100MB saved in memory usage in our bench.

#### master

|                      Method |        Job | UnrollFactor |        Mean |     Error |    StdDev |      Median |         Min |         Max |       Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|---------------------------- |----------- |------------- |------------:|----------:|----------:|------------:|------------:|------------:|------------:|-----------:|----------:|----------:|
|       CompileMethodsAndEmit | Job-ZZBXBK |           16 |  5,243.7 ms | 103.96 ms | 289.79 ms |  5,175.5 ms |  4,906.4 ms |  6,126.8 ms |  82000.0000 | 23000.0000 |         - |    497 MB |
|           SerializeMetadata | Job-ZZBXBK |           16 |    320.2 ms |   3.17 ms |   5.64 ms |    319.3 ms |    308.5 ms |    334.6 ms |   4000.0000 |  1000.0000 |         - |     34 MB |
|              GetDiagnostics | Job-KZXNEV |            1 |  4,314.1 ms |  97.91 ms | 287.16 ms |  4,209.7 ms |  3,901.5 ms |  5,160.1 ms |  73000.0000 | 21000.0000 |         - |    379 MB |
| GetDiagnosticsWithAnalyzers | Job-KZXNEV |            1 | 11,486.9 ms | 112.89 ms | 120.79 ms | 11,496.5 ms | 11,295.6 ms | 11,695.0 ms | 202000.0000 | 59000.0000 | 1000.0000 |  1,215 MB |

#### old-children

|                      Method |        Job | UnrollFactor |        Mean |    Error |    StdDev |      Median |         Min |         Max |       Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|---------------------------- |----------- |------------- |------------:|---------:|----------:|------------:|------------:|------------:|------------:|-----------:|----------:|----------:|
|       CompileMethodsAndEmit | Job-XXOYTR |           16 |  5,047.9 ms | 34.00 ms |  31.81 ms |  5,040.0 ms |  4,992.4 ms |  5,097.3 ms |  82000.0000 | 24000.0000 |         - |    497 MB |
|           SerializeMetadata | Job-XXOYTR |           16 |    325.8 ms |  2.97 ms |   5.13 ms |    324.6 ms |    317.4 ms |    338.2 ms |   3000.0000 |  1000.0000 |         - |     34 MB |
|              GetDiagnostics | Job-QCCYOV |            1 |  3,708.4 ms | 29.22 ms |  24.40 ms |  3,704.9 ms |  3,669.2 ms |  3,765.8 ms |  73000.0000 | 17000.0000 |         - |    378 MB |
| GetDiagnosticsWithAnalyzers | Job-QCCYOV |            1 | 11,112.4 ms | 99.31 ms | 135.93 ms | 11,096.5 ms | 10,884.7 ms | 11,436.8 ms | 195000.0000 | 55000.0000 | 1000.0000 |  1,178 MB |

#### better-children

|                      Method |        Job | UnrollFactor |        Mean |     Error |    StdDev |      Median |         Min |         Max |       Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|---------------------------- |----------- |------------- |------------:|----------:|----------:|------------:|------------:|------------:|------------:|-----------:|----------:|----------:|
|       CompileMethodsAndEmit | Job-UBJCRB |           16 |  5,113.7 ms |  40.92 ms |  34.17 ms |  5,097.6 ms |  5,078.3 ms |  5,198.1 ms |  81000.0000 | 22000.0000 |         - |    497 MB |
|           SerializeMetadata | Job-UBJCRB |           16 |    336.1 ms |   3.67 ms |  10.71 ms |    335.6 ms |    315.6 ms |    365.7 ms |   4000.0000 |  1000.0000 |         - |     34 MB |
|              GetDiagnostics | Job-OIOLXP |            1 |  3,768.2 ms |  23.62 ms |  20.94 ms |  3,765.4 ms |  3,741.8 ms |  3,821.4 ms |  72000.0000 | 19000.0000 |         - |    378 MB |
| GetDiagnosticsWithAnalyzers | Job-OIOLXP |            1 | 11,328.5 ms | 113.10 ms | 121.01 ms | 11,317.0 ms | 11,155.9 ms | 11,573.5 ms | 193000.0000 | 60000.0000 | 2000.0000 |  1,156 MB |

### RPS

A test insertion with this payload is [here](https://devdiv.visualstudio.com/DevDiv/_git/VS/pullrequest/291438). The DDRIT failures are unrelated, and a fix is being merged into dotnet/roslyn shortly.