@333fred 2020-12-07
Commit 1 is the only change that needs to be brought over from https://github.com/dotnet/roslyn/pull/49534. The issue itself exists in master, but is verified in this branch because of the additional Debug.Asserts, and was then triggered by AnalyzerRunner.

Commit 2 is the actual implementation. I used a struct enumerator with a slot-based approach: every `IOperation` or `ImmutableArray<IOperation>` has a slot. I also track the current index for the arrays as appropriate. Iterating advances through the slots and indexes. I've also updated our extension methods and walkers to use the new method, rather than `Children`, leading to the memory improvements detailed above.