@333fred 2020-12-07
@dotnet/roslyn-compiler @AlekseyTs @mavasani @sharwell please take a look.