@333fred 2020-12-09
@dotnet/roslyn-compiler @AlekseyTs for reviews please.