@AlekseyTs 2020-12-09
>                 0 when index < Operations.Length => Operations[index],

This is really unreadable, IMHO. Consider either moving the ```=> ...``` part on a new line, or using a regular switch statement. I believe that dense code is bad for readability. 

---
Refers to: src/Compilers/Core/Portable/Generated/Operations.Generated.cs:3335 in 10fdb5c. [](commit_id = 10fdb5c272ab20eb527dc04997ce94518bb4ad35, deletion_comment = False)