@AlekseyTs 2020-12-09
>                     return (false, int.MinValue, int.MinValue);

It looks like an unexpected ```previousSlot``` value is treated as the end of sequence. Should we be able to distinguish the two and, perhaps, throw on an invalid input.  

---
Refers to: src/Compilers/Core/Portable/Generated/Operations.Generated.cs:3348 in 10fdb5c. [](commit_id = 10fdb5c272ab20eb527dc04997ce94518bb4ad35, deletion_comment = False)