@AlekseyTs 2020-12-09
I think changes around the ```binder``` do not belong in this PR. Please separate them into separate PR with appropriate title and description about the reason for the change, etc.