@333fred 2020-12-10
>                     return (false, int.MinValue, int.MinValue);

> Should we be able to distinguish the two and, perhaps, throw on an invalid input

We could. I didn't think that this would particularly help us catch any bugs, but if you think it's valuable I can take that approach.

---
In reply to: [741991994](https://github.com/dotnet/roslyn/pull/49804#issuecomment-741991994) [](ancestors = 741991994)

---
Refers to: src/Compilers/Core/Portable/Generated/Operations.Generated.cs:3348 in 10fdb5c. [](commit_id = 10fdb5c272ab20eb527dc04997ce94518bb4ad35, deletion_comment = False)