@AlekseyTs 2020-12-10
>                     return (false, int.MinValue, int.MinValue);

> We could. I didn't think that this would particularly help us catch any bugs, but if you think it's valuable I can take that approach.

Explicitly transitioning to a dedicated "eof" slot (#1 in this case) and staying there might be clearer than to jumping to default, especially with ```goto default```. And the code generator could be somewhat simplified. 


---
In reply to: [742151554](https://github.com/dotnet/roslyn/pull/49804#issuecomment-742151554) [](ancestors = 742151554,741991994)

---
Refers to: src/Compilers/Core/Portable/Generated/Operations.Generated.cs:3348 in 10fdb5c. [](commit_id = 10fdb5c272ab20eb527dc04997ce94518bb4ad35, deletion_comment = False)