@333fred 2020-12-10
@AlekseyTs addressed feedback. I have an email out to the compat council for the breaking change.

---
In reply to: [742085337](https://github.com/dotnet/roslyn/pull/49804#issuecomment-742085337) [](ancestors = 742085337)