@333fred 2020-12-10
Other PR is merged. I'll wait till review on this is done, then rebase before merging.

---
In reply to: [742084981](https://github.com/dotnet/roslyn/pull/49804#issuecomment-742084981) [](ancestors = 742084981)