@AlekseyTs 2020-12-11
>             }

To clarify my comment in the class writer. I suggest to generate this here:
-Option 1:
```
                case 0:
                    return (false, 0, Operations.Length);
                default:
                    throw ExceptionUtilities.UnexpectedValue((previousSlot, previousIndex));
```

-Option 2:
```
                case 0:
                    goto case 1;
                case1:
                    return (false, 1, 0);
                default:
                    throw ExceptionUtilities.UnexpectedValue((previousSlot, previousIndex));
```

---
Refers to: src/Compilers/Core/Portable/Generated/Operations.Generated.cs:3352 in 0b096a1. [](commit_id = 0b096a10f54e1e2b7e32a3ffb825458c024712eb, deletion_comment = False)