@AlekseyTs 2020-12-11
>             }

For this case, the last property is not an array, I suggest this code generation:
```
case 0:
    goto case 1;
case1:
    return (false, 1, 0);
default:
    throw ExceptionUtilities.UnexpectedValue((previousSlot, previousIndex));
```

---
Refers to: src/Compilers/Core/Portable/Generated/Operations.Generated.cs:3698 in 0b096a1. [](commit_id = 0b096a10f54e1e2b7e32a3ffb825458c024712eb, deletion_comment = False)