@333fred 2020-12-11
>             }

Yep, this is about what I was thinking, expect that I didn't have a `goto case 1;`, I just let it fall through.

---
In reply to: [743363000](https://github.com/dotnet/roslyn/pull/49804#issuecomment-743363000) [](ancestors = 743363000)

---
Refers to: src/Compilers/Core/Portable/Generated/Operations.Generated.cs:3698 in 0b096a1. [](commit_id = 0b096a10f54e1e2b7e32a3ffb825458c024712eb, deletion_comment = False)