@AlekseyTs 2020-12-11
>             }

> I just let it fall through.

That works too.

---
In reply to: [743366535](https://github.com/dotnet/roslyn/pull/49804#issuecomment-743366535) [](ancestors = 743366535,743363000)

---
Refers to: src/Compilers/Core/Portable/Generated/Operations.Generated.cs:3698 in 10fdb5c. [](commit_id = 10fdb5c272ab20eb527dc04997ce94518bb4ad35, deletion_comment = False)