@333fred 2020-12-11
>             }

Went with option 2.

---
In reply to: [743358865](https://github.com/dotnet/roslyn/pull/49804#issuecomment-743358865) [](ancestors = 743358865)

---
Refers to: src/Compilers/Core/Portable/Generated/Operations.Generated.cs:3352 in 0b096a1. [](commit_id = 0b096a10f54e1e2b7e32a3ffb825458c024712eb, deletion_comment = False)