@333fred 2020-12-11
@dotnet/roslyn-compiler for a second review.